# task2/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Gateway for GitHub repository information",
        "title": "GitHub Repository API Gateway",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/repos/{owner}/{repo}": {
            "get": {
                "description": "Get information about a GitHub repository",
                "tags": [
                    "repositories"
                ],
                "summary": "Get repository info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository owner",
                        "name": "owner",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "repo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Repository not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Upstream error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}